# docs/openapi-doc.yml
openapi: 3.1.0
info:
    title: Firepit API
    version: 1.7.0
    description: |
        Firepit is a realtime communication platform built with Next.js 16, React 19, Bun, and Appwrite.

        This specification covers the public and first-party client API surface currently used by the application.
        Internal debug and test-only endpoints are intentionally omitted.
    contact:
        name: Firepit
        url: https://github.com/acarlson33/firepit
    license:
        name: GNU General Public License v3.0
        url: https://opensource.org/licenses/GPL-3.0
servers:
    - url: http://localhost:3000
      description: Local development server
    - url: https://firepit.example.com
      description: Example production server
...
components:
    securitySchemes:
        sessionCookie:
            type: apiKey
            in: cookie
            name: a_session_*
            description: Appwrite session cookie used by authenticated first-party clients
    schemas:
...
        Success:
            type: object
            properties:
                success:
                    type: boolean
                    default: true
...
paths:
...
    /api/upload-emoji:
        post:
            tags: [Uploads, Emojis]
            summary: Upload a custom emoji asset
            security:
                - sessionCookie: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
                                name:
                                    type: string
            responses:
                "200":
                    description: Emoji uploaded
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: true
        delete:
            tags: [Uploads, Emojis]
            summary: Delete a custom emoji asset
            security:
                - sessionCookie: []
            parameters:
                - name: fileId
                  in: query
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Emoji deleted
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Success"
    /api/emoji/{fileId}:
        get:
            tags: [Emojis]
            summary: Proxy or retrieve a custom emoji asset
            security: []
            parameters:
                - name: fileId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Emoji asset stream
                    content:
                        image/*:
                            schema:
                                type: string
                                format: binary
...